service/requirements.txt
defusedxml==0.4.1
Django==1.10.2
mysqlclient==1.3.9
oauthlib==2.0.0
PyJWT==1.4.2
python-social-auth==0.2.21
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
wheel==0.24.0
